Migrated func _verify_sign_in_token to BitwardenPasswordlessClient
Removed func `backend._verify_sign_in_token` of `bitwarden_passwordless`
sub-package and moved its functionality to the method
`client.BitwardenPasswordlessClient.verify_sign_in`, since the
functionality of the `backend` module can be moved to the `client` module
now that the `client` module only handles backend operations and no frontend
operations. Migrated unit tests for `backend._verify_sign_in_token`.
