+++ src/streamlit_passwordless/bitwarden_passwordless/backend.py
@@ -6,13 +6,11 @@ from datetime import datetime, timedelta
 from typing import Literal, Self, TypeAlias
 
 # Third party
-from passwordless import PasswordlessClient, PasswordlessError, VerifiedUser, VerifySignIn
+from passwordless import PasswordlessClient, VerifiedUser
 from pydantic import AnyHttpUrl
 
 # Local
-from streamlit_passwordless import common, exceptions, models
-
-BackendClient: TypeAlias = PasswordlessClient
+from streamlit_passwordless import common, models
 
 logger = logging.getLogger(__name__)
 
@@ -149,46 +147,3 @@ class BitwardenPasswordlessVerifiedUser(models.BaseModel):
             type=verified_user.type,
             rp_id=verified_user.rp_id,
         )
-
-
-def _verify_sign_in_token(client: BackendClient, token: str) -> BitwardenPasswordlessVerifiedUser:
-    r"""Verify the sign in token to complete the sign in process.
-
-    The sign in token is generated from the `frontend._sign_in` function.
-
-    Parameters
-    ----------
-    client : BackendClient
-        The Bitwarden Passwordless backend client to communicate
-        with the Bitwarden Passwordless backend.
-
-    token : str
-        The token to verify.
-
-    Returns
-    -------
-    BitwardenPasswordlessVerifiedUser
-        Details from Bitwarden Passwordless about the user that was signed in.
-
-    Raises
-    ------
-    streamlit_passwordless.SignInTokenVerificationError
-        If the `token` cannot be verified successfully.
-
-    streamlit_passwordless.StreamlitPasswordlessError
-        If an instance of `BitwardenPasswordlessVerifiedUser` cannot be successfully created.
-    """
-
-    try:
-        _verified_user = client.sign_in(verify_sign_in=VerifySignIn(token=token))
-    except PasswordlessError as e:
-        error_msg = f'Error verifying the sign in token!\nproblem_details: {e.problem_details}'
-        data = {
-            'token': token,
-            'problem_details': e.problem_details,
-        }
-        raise exceptions.SignInTokenVerificationError(error_msg, data=data) from None
-
-    return BitwardenPasswordlessVerifiedUser._from_passwordless_verified_user(
-        verified_user=_verified_user
-    )
